# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/properties": {
      "get": {
        "operationId": "AdminPropertiesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "pending_review",
              "enum": [
                "draft",
                "pending_review",
                "needs_revision",
                "approved",
                "rejected",
                "suspended",
                "escalated"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List properties in the moderation queue",
        "tags": [
          "Admin / Properties"
        ]
      }
    },
    "/admin/properties/{propertyId}": {
      "get": {
        "operationId": "AdminPropertiesController_getDetail",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full property detail for moderation",
        "tags": [
          "Admin / Properties"
        ]
      }
    },
    "/admin/properties/{propertyId}/approve": {
      "post": {
        "operationId": "AdminPropertiesController_approve",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve a pending property submission",
        "tags": [
          "Admin / Properties"
        ]
      }
    },
    "/admin/properties/{propertyId}/reject": {
      "post": {
        "operationId": "AdminPropertiesController_reject",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a pending property submission",
        "tags": [
          "Admin / Properties"
        ]
      }
    },
    "/admin/properties/{propertyId}/request-revision": {
      "post": {
        "operationId": "AdminPropertiesController_requestRevision",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRevisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request changes to a pending property submission",
        "tags": [
          "Admin / Properties"
        ]
      }
    },
...
    "/owners/me/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [
          {
            "name": "unread",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the caller's notification inbox",
        "tags": [
          "Owner Notifications"
        ]
      }
    },
    "/owners/me/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a notification as read",
        "tags": [
          "Owner Notifications"
        ]
      }
    },
...
    "/bookings": {
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new booking (pending_payment)",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full booking detail",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/payment/confirm": {
      "post": {
        "operationId": "BookingsController_confirmPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mock payment gateway confirmation",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/check-in": {
      "post": {
        "operationId": "BookingsController_checkIn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check in using the booking QR code",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/check-out": {
      "post": {
        "operationId": "BookingsController_checkOut",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually check out a booking",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "post": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a confirmed booking",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/admin/bookings/{id}": {
      "get": {
        "operationId": "AdminBookingsController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Full booking detail incl. timeline and admin actions",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
...
    "/admin/bookings/{id}/cancel": {
      "post": {
        "operationId": "AdminBookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCancelBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin-initiated cancellation",
        "tags": [
          "Admin / Bookings"
        ]
      }
    },
...
  },
  "info": {
    "title": "QuickNest API",
    "description": "Backend APIs for QuickNest hotel booking platform",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RejectPropertyDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "reason"
        ]
      },
      "RevisionItemDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ]
      },
      "RequestRevisionDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevisionItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string",
            "description": "Property being booked"
          },
          "roomTypeId": {
            "type": "string",
            "description": "Room type within the property"
          },
          "bookingType": {
            "type": "string",
            "enum": [
              "hourly",
              "fullday"
            ]
          },
          "checkInAt": {
            "type": "string",
            "description": "Requested check-in instant, ISO timestamp"
          },
          "durationHours": {
            "type": "number",
            "description": "Required for hourly bookings; ignored for fullday (fixed at 24)"
          },
          "guestCount": {
            "type": "number",
            "description": "Number of guests for this booking"
          }
        },
        "required": [
          "propertyId",
          "roomTypeId",
          "bookingType",
          "checkInAt",
          "guestCount"
        ]
      },
      "PaymentConfirmDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Mock payment gateway result"
          },
          "paymentRef": {
            "type": "string",
            "description": "Mock gateway transaction id"
          }
        },
        "required": [
          "success"
        ]
      },
      "CheckInDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "QR code token issued on booking confirmation"
          }
        },
        "required": [
          "qrCode"
        ]
      },
      "CancelBookingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Guest-supplied cancellation reason"
          }
        }
      },
...
      "AdminCancelBookingDto": {
        "type": "object",
        "properties": {
          "reasonCategory": {
            "type": "string",
            "enum": [
              "guest_request",
              "hotel_request",
              "admin_decision",
              "system_error"
            ]
          },
          "reasonText": {
            "type": "string"
          }
        },
        "required": [
          "reasonCategory"
        ]
      },
...
    }
  }
}